Modified the protocol.
Modified the protocol to support a CRC flag indicating if a CRC32 is part of the header; permit peer ids greater than 4095; restrict channel count to a maximum of 256 channels and maximum fragment count to 65000. Fixed warnings about integer type conversions and a possible DOS exploit due to the lack of proper validation of packet data lengths in certain cases. Added a few extra functions to improve the C# wrapper. Added a test project.